Increased string size, moved PathData
