+++ Autopilot/Path Manager/Utilities/NMEAParser.c
@@ -7,6 +7,7 @@
  */
 
 #include "../../Common/Common.h"
+#include "../../Common/Utilities/Logger.h"
 #include <stdbool.h>
 #include <ctype.h>
 #include <stdio.h>
@@ -84,7 +85,7 @@ void parseGGA(char* data, long double* latitude, long double* longitude, float* 
 }
 
 void parseVTG(char* data, float* speed, int16_t* heading) {
-	char values[10][7]; //store the value of the 10 components of this message (overkill, only 2 are needed)
+	char values[20][7]; //store the value of the 10 components of this message (overkill, only 2 are needed)
 	int i = 0;
 	int j = 0;
 	int n = 0;
@@ -99,12 +100,11 @@ void parseVTG(char* data, float* speed, int16_t* heading) {
 		}
 		j++;
 	}
-
 	// fill the GPSData struct with info from values[]
 	sscanf(values[7], "%f", speed);
-        if(values[1] != 0) sscanf(values[1], "%d", heading);
-
-	*speed = *speed / 3.6;
+    if(values[1] != 0) sscanf(values[1], "%d", heading);
+    
+    *speed = *speed / 3.6;
 }
 
 /**
